requirements.txt
argparse
numpy
python-osc
tqdm
pydub
praat-parselmouth
unidecode
